うんこinけつもう

+++ src/main/java/org/meyason/dokkoi/event/player/DamageEvent.java
@@ -1,16 +1,6 @@
 package org.meyason.dokkoi.event.player;
 
-import org.bukkit.Location;
-import org.bukkit.NamespacedKey;
-import org.bukkit.Particle;
-import org.bukkit.Sound;
-import org.bukkit.entity.*;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageByEntityEvent;
-import org.bukkit.persistence.PersistentDataType;
-import org.bukkit.projectiles.ProjectileSource;
-import org.bukkit.util.Vector;
+import net.kyori.adventure.sound.Sound;
 import org.meyason.dokkoi.Dokkoi;
 import org.meyason.dokkoi.constants.EntityID;
 import org.meyason.dokkoi.constants.GameItemKeyString;
@@ -24,21 +14,52 @@ import org.meyason.dokkoi.game.ProjectileData;
 import org.meyason.dokkoi.item.job.Rapier;
 import org.meyason.dokkoi.job.*;
 
-import java.util.List;
+import org.spongepowered.api.data.Keys;
+import org.spongepowered.api.data.type.PickupRules;
+import org.spongepowered.api.effect.VanishState;
+import org.spongepowered.api.effect.particle.ParticleEffect;
+import org.spongepowered.api.effect.particle.ParticleTypes;
+import org.spongepowered.api.effect.sound.SoundType;
+import org.spongepowered.api.effect.sound.SoundTypes;
+import org.spongepowered.api.entity.Entity;
+import org.spongepowered.api.entity.living.Living;
+import org.spongepowered.api.entity.living.player.Player;
+import org.spongepowered.api.entity.living.trader.Villager;
+import org.spongepowered.api.entity.projectile.Projectile;
+import org.spongepowered.api.entity.projectile.Snowball;
+import org.spongepowered.api.entity.projectile.arrow.Arrow;
+import org.spongepowered.api.entity.projectile.arrow.Trident;
+import org.spongepowered.api.event.Cause;
+import org.spongepowered.api.event.EventListener;
+import org.spongepowered.api.event.entity.DamageEntityEvent;
+import org.spongepowered.api.projectile.source.ProjectileSource;
 
-public class DamageEvent implements Listener {
+import java.util.List;
+import java.util.Optional;
+
+public class DamageEvent implements EventListener<DamageEntityEvent> {
+
+    // イベントのメインハンドラ
+    @Override
+    public void handle(DamageEntityEvent event){
+        Entity damaged = event.entity();
+        Cause cause = event.cause();
+
+        cause.first(Player.class).ifPresent(player -> {
+            if(damaged instanceof Player damagedPlayer){
+                fromPlayertoPlayerDamage(event, player, damagedPlayer);
+            }else{
+                fromPlayerToEntityDamage(event, player, damaged);
+            }
+        });
+
+        cause.first(Projectile.class).ifPresent(projectile -> {
+            fromProjectileDamage(event, projectile, damaged);
+        });
+    }
 
     // プレイヤーからプレイヤーへのダメージ
-    @EventHandler
-    public void onEntityDamageByEntity(EntityDamageByEntityEvent event){
-        Entity attacker = event.getDamager();
-        Entity damaged = event.getEntity();
-        if(!(attacker instanceof Player attackedPlayer) || !(damaged instanceof Player damagedPlayer)) return;
-
-        if(damagedPlayer.getNoDamageTicks() >= 10){
-            event.setCancelled(true);
-            return;
-        }
+    public void fromPlayertoPlayerDamage(DamageEntityEvent event, Player attackedPlayer, Player damagedPlayer) {
         Game game = Game.getInstance();
         GameStatesManager gameStatesManager = game.getGameStatesManager();
         if(gameStatesManager.getGameState() != GameState.IN_GAME) {
@@ -56,32 +77,30 @@ public class DamageEvent implements Listener {
         }
 
         event.setCancelled(true);
-        double damage = event.getFinalDamage();
-        calculateDamage(attackedPlayer, damagedPlayer, damage);
-
+        calculateDamage(attackedPlayer, damagedPlayer, event.finalDamage());
     }
 
-    // エンティティからのダメージ
-    @EventHandler
-    public void onPlayerHit(EntityDamageByEntityEvent event){
-        if(event.getDamager() instanceof Player) return;
-
+    // Projectileからのダメージ
+    public void fromProjectileDamage(DamageEntityEvent event, Projectile projectile, Entity damagedEntity) {
         GameStatesManager gameStatesManager = Game.getInstance().getGameStatesManager();
-        Entity entity = event.getEntity();
-        if (event.getEntity().isDead()) {
+        if (projectile.isRemoved()) {
             return;
         }
-        if (!(entity instanceof LivingEntity livingEntity)) {
+        if (!(damagedEntity instanceof Living livingEntity)) {
             return;
         }
 
         Player attackedPlayer = null;
         Player damagedPlayer = null;
-        Entity damagedEntity = event.getEntity();
-        double damage = event.getFinalDamage();
+        double damage = event.finalDamage();
+
+        Optional<ProjectileSource> shooterOpt = projectile.get(Keys.SHOOTER);
+        if (shooterOpt.isEmpty()) {
+            return;
+        }
 
         // だいたいのスキル系投擲物はsnowball
-        if(event.getDamager() instanceof Snowball snowball) {
+        if(projectile instanceof Snowball snowball) {
             ProjectileData projectileData = gameStatesManager.getProjectileDataMap().get(snowball);
             if (projectileData == null) {
                 return;
@@ -99,9 +118,9 @@ public class DamageEvent implements Listener {
                 String attackItem = projectileData.getItem();
                 if(attackItem.equals(GameItemKeyString.SKILL)) {
                     List<Player> effectedPlayers = CalculateAreaPlayers.getPlayersInArea(Game.getInstance(), attackedPlayer, snowball.getLocation(), 10);
-                    bomber.skill(snowball.getLocation(), effectedPlayers);
+                    bomber.skill(snowball.location(), effectedPlayers);
                 }else if(attackItem.equals(GameItemKeyString.ULTIMATE_SKILL)){
-                    bomber.ultimate(snowball.getLocation());
+                    bomber.ultimate(snowball.location());
                 }
                 gameStatesManager.removeProjectileData(snowball);
                 return;
@@ -133,13 +152,15 @@ public class DamageEvent implements Listener {
                 return;
             }
 
-        }else if(event.getDamager() instanceof Trident trident) {
+        }else if(projectile instanceof Trident trident) {
             ProjectileData projectileData = gameStatesManager.getProjectileDataMap().get(trident);
             // 特殊アイテムじゃないばあい(素のトライデント)
             if (projectileData == null) {
                 return;
             }
-            trident.setPickupStatus(AbstractArrow.PickupStatus.DISALLOWED);
+
+            trident.offer(Keys.PICKUP_RULE, PickupRules.DISALLOWED.get());
+
             if(projectileData.getItem().equals(Rapier.id)) {
                 attackedPlayer = projectileData.getAttacker();
                 if(gameStatesManager.getPlayerJobs().get(attackedPlayer) instanceof Lonely lonely){
@@ -151,39 +172,41 @@ public class DamageEvent implements Listener {
                     event.setCancelled(true);
 
                     Rapier rapier = ironMaiden.getRapier();
-                    rapier.activate(trident, trident.getLocation());
+                    rapier.activate(trident, trident.location());
                     gameStatesManager.removeProjectileData(trident);
                     return;
                 }
             }
 
-        }else if(event.getDamager() instanceof Arrow arrow) {
+        }else if(projectile instanceof Arrow arrow) {
             ProjectileData projectileData = gameStatesManager.getProjectileDataMap().get(arrow);
 
             if (projectileData == null) {
-                ProjectileSource source = arrow.getShooter();
-                if(source instanceof Entity shooterEntity){
-                    if(shooterEntity instanceof Player attackerPlayer){
-                        if(gameStatesManager.getPlayerJobs().get(attackerPlayer) instanceof Lonely lonely){
-                            lonely.lastAttackedTime = System.currentTimeMillis();
-                        }
+                Entity shooterEntity = (Entity) shooterOpt.get();
+                if (shooterEntity instanceof Player attackerPlayer) {
+                    if (gameStatesManager.getPlayerJobs().get(attackerPlayer) instanceof Lonely lonely) {
+                        lonely.lastAttackedTime = System.currentTimeMillis();
                     }
-                    if(damagedEntity instanceof Player damagedP){
-                        if(gameStatesManager.getPlayerJobs().get(damagedP) instanceof Lonely lonely){
-                            lonely.lastDamagedTime = System.currentTimeMillis();
-                        }
-                    }
-                    calculateDamage(shooterEntity, damagedEntity, damage);
                 }
+                if(damagedEntity instanceof Player damagedP){
+                    if(gameStatesManager.getPlayerJobs().get(damagedP) instanceof Lonely lonely){
+                        lonely.lastDamagedTime = System.currentTimeMillis();
+                    }
+                }
+                calculateDamage(shooterEntity, damagedEntity, damage);
                 return;
             }
 
             attackedPlayer = projectileData.getAttacker();
             if(gameStatesManager.getPlayerJobs().get(attackedPlayer) instanceof Explorer) {
-                //自分が放つ矢が着弾した位置に爆発を起こす。爆発は当たった対象に固定10ダメージを与える。
-                arrow.getWorld().spawnParticle(Particle.EXPLOSION, arrow.getLocation(), 1);
-                arrow.getWorld().playSound(arrow.getLocation(), Sound.ENTITY_GENERIC_EXPLODE, 10.0F, 1.0F);
-                List<Player> effectedPlayers = CalculateAreaPlayers.getPlayersInArea(Game.getInstance(), attackedPlayer, arrow.getLocation(), 3);
+                //自分が放つ矢が着弾した位置に爆発を起こす。爆発は当たった対象に固定10ダメージを与える
+                ParticleEffect explosionEffect = ParticleEffect.builder()
+                        .type(ParticleTypes.EXPLOSION_EMITTER)
+                        .build();
+                arrow.world().spawnParticles(explosionEffect, arrow.location().position());
+                Sound sound = Sound.sound(SoundTypes.ENTITY_GENERIC_EXPLODE, Sound.Source.MASTER, 10.0f, 1.0f);
+                arrow.world().playSound(sound, arrow.location().position());
+                List<Player> effectedPlayers = CalculateAreaPlayers.getPlayersInArea(Game.getInstance(), attackedPlayer, arrow.location(), 3);
                 effectedPlayers.add(attackedPlayer);
                 gameStatesManager.addAttackedPlayer(attackedPlayer);
                 for (Player damaged : effectedPlayers) {
@@ -197,18 +220,22 @@ public class DamageEvent implements Listener {
         }
 
         calculateDamage(attackedPlayer, damagedPlayer, damage);
-
     }
 
-    //　プレイヤーからエンティティへのダメージ
-    @EventHandler
-    public static void onEntityDamage(EntityDamageByEntityEvent event){
-        Entity damaged = event.getEntity();
-        if(damaged instanceof Player){return;}
-        if(!(event.getDamager() instanceof Player attacker)){return;}
-        if(damaged.isDead()){
+//    // エンティティからのダメージ
+//    public void fromEntityDamage(DamageEntityEvent event, Entity attacker, Player damagedPlayer) {
+//
+//    }
+
+    // プレイヤーからエンティティへのダメージ
+    public void fromPlayerToEntityDamage(DamageEntityEvent event, Player attacker, Entity damagedEntity) {
+        if(damagedEntity instanceof Player){return;}
+        Optional<VanishState> vanishStateOpt = attacker.get(Keys.VANISH_STATE);
+        if(vanishStateOpt.isPresent() && vanishStateOpt.get().equals(VanishState.vanished())){
+            event.setCancelled(true);
             return;
         }
+
         Game game = Game.getInstance();
         GameStatesManager gameStatesManager = game.getGameStatesManager();
         if(gameStatesManager.getGameState() != GameState.IN_GAME) {
@@ -220,14 +247,10 @@ public class DamageEvent implements Listener {
             lonely.lastAttackedTime = System.currentTimeMillis();
         }
 
-        if(damaged instanceof Villager villager){
-            for(EntityID entityID : EntityID.values()){
-                if(villager.getPersistentDataContainer().has(new NamespacedKey(Dokkoi.getInstance(), entityID.getId()), PersistentDataType.STRING)){
-                    if(Comedian.isComedianByID(entityID.getId())){
-                        event.setCancelled(true);
-                        return;
-                    }
-                }
+        if(damagedEntity instanceof Villager villager){
+            if(gameStatesManager.isComedian(villager)){
+                event.setCancelled(true);
+                return;
             }
         }
     }
@@ -257,10 +280,17 @@ public class DamageEvent implements Listener {
                 return;
             }
 
-            double afterHealth = damagedPlayer.getHealth() - damage;
+            Optional<Double> optHealth = damagedPlayer.get(Keys.HEALTH);
+            if (optHealth.isEmpty()) {
+                return;
+            }
+            double currentHealth = optHealth.get();
+            double afterHealth = currentHealth - damage;
             // 死亡処理
             if (afterHealth <= 0) {
                 DeathEvent.kill(attackerPlayer, damagedPlayer);
+            }else{
+                damagedPlayer.offer(Keys.HEALTH, afterHealth);
             }
         }
     }
